рефакторинг кода, удален класс ErrorResponse

+++ src/main/java/ru/yandex/practicum/filmorate/storage/FilmStorage.java
@@ -13,7 +13,7 @@ public interface FilmStorage {
 
     void delete(Film film);
 
-    Integer getSize();
+    long getSize();
 
     Collection<Film> getFilms();
