+++ src/main/java/ru/yandex/practicum/filmorate/storage/UserStorage.java
@@ -12,7 +12,7 @@ public interface UserStorage {
 
     void delete(User user);
 
-    Integer getSize();
+    long getSize();
 
     Collection<User> getUsers();
